Added the napari-plugin files
So I tried making a napari plugin and ended up just disabling some of the widgets to oinly keep the QtWidgets as they seem more intuitive. It is still, in functionality, the same as the base example plugin that is given. I am trying to figure out how to use the "viewer.xxx" variable. I'm not sure where it comes from (it's supposed to be what the viewer of Napari is that displays the images so it gives access to the layers) and I could not yet find documentation that describes what else is available from "viewer.xxx" since I am trying to use the mask layer from CellPose2. I will most probably need to do this in a project that lets me program without having to restart Napari each time.